Add a max sleep interval
If the waiting time is too long, such as seconds, minutes, etc., it will not be possible to cancel the job immediately, so instead of waiting for the delta time between the rest interval, add a maximum rest value.
If the value is negative, the waiting time will be the waiting time until the next tick, a positive value will be taken as the waiting value.